# src/main/resources/META-INF/backbone_rest-openapi.yaml
openapi: "3.1.0"
info:
  title: "Backbone REST API"
  description: "backbone_rest API"
  version: "1.0.0"
servers:
- url: "https://backbone_rest"
...
paths:
  /api/v1/contacts/{contactId}:
    put:
      summary: ""
      description: "Update a contact."
      operationId: ""
      parameters:
      - name: "contactId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contact"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      summary: ""
      description: "Delete a contact."
      operationId: ""
      parameters:
      - name: "contactId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
    get:
      summary: ""
      description: "Find a contact list."
      operationId: ""
      parameters:
      - name: "contactId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /api/v1/contacts/:
    post:
      summary: ""
      description: "Create a contact."
      operationId: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contact"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /api/v1/contacts/list/{contactIds}:
    get:
      summary: ""
      description: "Find a contact list by ids."
      operationId: ""
      parameters:
      - name: "contactIds"
        in: "path"
        required: true
        schema:
          type: "array"
          items:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Contact"
  /api/v1/contacts/person/{personId}:
    get:
      summary: ""
      description: "List contacts by person Id."
      operationId: ""
      parameters:
      - name: "personId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Contact"
...
  /api/v1/features/{featureId}:
    put:
      summary: ""
      description: "Update a Feature"
      operationId: ""
      parameters:
      - name: "featureId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
  /api/v1/features/:
    post:
      summary: ""
      description: "Create a Feature"
      operationId: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
  /api/v1/features/find/{featureId}:
    get:
      summary: ""
      description: "Realiza la busqueda da un feature"
      operationId: ""
      parameters:
      - name: "featureId"
        in: "path"
        required: true
        description: "Id de feature"
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
  /api/v1/features/{includeInactive}:
    get:
      summary: ""
      description: "Get a list of active and/or inactive features."
      operationId: ""
      parameters:
      - name: "includeInactive"
        in: "path"
        required: true
        description: "Non/Include the features inactive"
        schema:
          type: "boolean"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Feature"
  /api/v1/features/{includeInactive}/{featuresIds}:
    get:
      summary: ""
      description: "Get a list of active and/or inactive features."
      operationId: ""
      parameters:
      - name: "includeInactive"
        in: "path"
        required: true
        description: "Non/Include the features inactive"
        schema:
          type: "boolean"
      - name: "featuresIds"
        in: "path"
        required: true
        description: "Features id list"
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Feature"
...
  /api/v1/contact-types/{contactTypeId}:
    put:
      summary: ""
      description: "Find a contact type list."
      operationId: ""
      parameters:
      - name: "contactTypeId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactType"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactType"
    delete:
      summary: ""
      description: "Delete a contact type."
      operationId: ""
      parameters:
      - name: "contactTypeId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactType"
    get:
      summary: ""
      description: "Find a contact type list."
      operationId: ""
      parameters:
      - name: "contactTypeId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactType"
  /api/v1/contact-types/:
    post:
      summary: ""
      description: "Create a contact type."
      operationId: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactTypeRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactType"
  /api/v1/contact-types/list/{contactTypeIds}:
    get:
      summary: ""
      description: "Find a contact type list by ids."
      operationId: ""
      parameters:
      - name: "contactTypeIds"
        in: "path"
        required: true
        schema:
          type: "array"
          items:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ContactType"
...
components:
  schemas:
    ContactType:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: "string"
        active:
          type: "boolean"
    Person:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        firstName:
          type: "string"
        middleName:
          type: "string"
        lastName:
          type: "string"
        gender:
          type: "string"
        birthdate:
          type: "string"
          format: "date"
    Contact:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        content:
          type: "string"
        contactType:
          $ref: "#/components/schemas/ContactType"
        person:
          $ref: "#/components/schemas/Person"
        active:
          type: "boolean"
    Feature:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: "string"
        active:
          type: "boolean"
    FeatureRequest:
      type: "object"
      properties:
        feature:
          $ref: "#/components/schemas/Feature"
        dateTime:
          type: "string"
          format: "date-time"
        appName:
          type: "string"
        appToken:
          type: "string"
...
    ContactTypeRequest:
      type: "object"
      properties:
        contactType:
          $ref: "#/components/schemas/ContactType"
        dateTime:
          type: "string"
          format: "date-time"
        appName:
          type: "string"
        appToken:
          type: "string"